Fix bug in parsing partial message headers
The problem is that if a message fragment ends with a complete header,
that header is effectively skipped, because the "bodyOffset" variable
is advanced past that last header.

Only advance past the last header if AdvanceAndUnfold was successful.
Otherwise, don't advance past it, and expect that same header to be
passed in next time.